+++ include/moving_sphere.h
@@ -10,30 +10,30 @@ class moving_sphere : public hittable {
 public:
     moving_sphere() {}
     moving_sphere(
-        point3 cen0, point3 cen1, double _time0, double _time1, double r, shared_ptr<material> m)
+        point3 cen0, point3 cen1, Float _time0, Float _time1, Float r, shared_ptr<material> m)
         : center0(cen0), center1(cen1), time0(_time0), time1(_time1), radius(r), mat_ptr(m)
     {};
 
     virtual bool hit(
-        const ray& r, double t_min, double t_max, hit_record& rec) const override;
+        const ray& r, Float t_min, Float t_max, hit_record& rec) const override;
 
     virtual bool bounding_box(
-        double _time0, double _time1, aabb& output_box) const override;
+        Float _time0, Float _time1, aabb& output_box) const override;
 
-    point3 center(double time) const;
+    point3 center(Float time) const;
 
 public:
     point3 center0, center1;
-    double time0, time1;
-    double radius;
+    Float time0, time1;
+    Float radius;
     shared_ptr<material> mat_ptr;
 };
 
-point3 moving_sphere::center(double time) const {
+point3 moving_sphere::center(Float time) const {
     return center0 + ((time - time0) / (time1 - time0)) * (center1 - center0);
 }
 
-bool moving_sphere::hit(const ray& r, double t_min, double t_max, hit_record& rec) const {
+bool moving_sphere::hit(const ray& r, Float t_min, Float t_max, hit_record& rec) const {
     vec3 oc = r.origin() - center(r.time());
     auto a = r.direction().length_squared();
     auto half_b = dot(oc, r.direction());
@@ -60,7 +60,7 @@ bool moving_sphere::hit(const ray& r, double t_min, double t_max, hit_record& re
     return true;
 }
 
-bool moving_sphere::bounding_box(double _time0, double _time1, aabb& output_box) const {
+bool moving_sphere::bounding_box(Float _time0, Float _time1, aabb& output_box) const {
     aabb box0(
         center(_time0) - vec3(radius, radius, radius),
         center(_time0) + vec3(radius, radius, radius));
