+++ include/onb.h
@@ -12,7 +12,7 @@ public:
     vec3 v() const { return axis[1]; }
     vec3 w() const { return axis[2]; }
 
-    vec3 local(double a, double b, double c) const {
+    vec3 local(Float a, Float b, Float c) const {
         return a * u() + b * v() + c * w();
     }
