+++ include/ray.h
@@ -6,22 +6,22 @@
 class ray {
 public:
     ray() {}
-    ray(const point3& origin, const vec3& direction, double time = 0.0)
+    ray(const point3& origin, const vec3& direction, Float time = 0.0)
         : orig(origin), dir(direction), tm(time)
     {}
 
     point3 origin() const { return orig; }
     vec3 direction() const { return dir; }
-    double time() const { return tm; }
+    Float time() const { return tm; }
 
-    point3 at(double t) const {
+    point3 at(Float t) const {
         return orig + t * dir;
     }
 
 public:
     point3 orig;
     vec3 dir;
-    double tm;
+    Float tm;
 };
 
 #endif
